Completed the Caesar Cipher by first creating an alphabet array.
I created a current number variable that was used to test whether a number would hold the current value and test whether the current letter + 13 would be outside of the array.

The current number was taken from the string that's passed to the function using the alphabet array and indexOf.

Another thing I needed to test for were things that aren't in the alphabet array (symbols, numbers, etc.). If current returned an index of -1, I just inserted the value of str[i] (which would be a symbol or whatever).

If the number was >= 13 I subtracted 13 from it and added the value to the decoded variable.

If the number was something other than -1 or >= 13, then it added 13 to the current value and added it to the decoded variable.